fixing inconsistencies

+++ src/AuthenticationServer.API/Models/AuthConfiguration.cs
@@ -3,9 +3,9 @@
     public class AuthConfiguration
     {
         public string AccessTokenSecretKey { get; set; }
-        public double AccessTokenExpirationMinutes { get; set; }
+        public int AccessTokenExpirationMinutes { get; set; }
         public string RefreshTokenSecretKey { get; set; }
-        public double RefreshTokenExpirationMinutes { get; set; }
+        public int RefreshTokenExpirationMinutes { get; set; }
         public string Issuer { get; set; }
         public string Audience { get; set; }
     }
