+++ src/Example.API/Startup.cs
@@ -71,7 +71,7 @@ namespace Example.API
         private class AuthConfiguration
         {
             public string AccessTokenSecretKey { get; set; }
-            public double AccessTokenExpirationMinutes { get; set; }
+            public int AccessTokenExpirationMinutes { get; set; }
             public string Issuer { get; set; }
             public string Audience { get; set; }
         }
